Added initial bio for GitHub profile
This commit adds the initial bio for my GitHub profile. It highlights my interest in leveraging data for insights, ongoing learning in Python, SQL, Power BI, and Excel for Data Science, and seeking collaboration opportunities in data analysis and visualization projects, along with contact information. Additionally, it includes my preferred pronouns and a fun fact about my passion for exploring datasets. This bio aims to provide a comprehensive overview of my professional interests and background in the field of Data Science.